abstracted styles out in some forms
