+++ frontend/src/components/SignIn.jsx
@@ -1,61 +1,14 @@
 import React, { useState } from "react";
-import styled from "styled-components";
-import Container from "@mui/material/Container";
 import { useNavigate } from "react-router-dom";
-
-const StyledBlock = styled(Container)`
-  padding: 100px;
-  margin: 75px;
-  justify-content: center;
-  background: #f6da73;
-  box-shadow: 5px 5px 5px gray;
-  width: 500px;
-`;
-
-const StyledHead = styled.div`
-  font-style: normal;
-  font-weight: 400;
-  font-size: 48px;
-  line-height: 58px;
-  text-align: center;
-`;
-
-const StyledText = styled.div`
-  font-style: normal;
-  font-size: 30px;
-  padding-top: 25px;
-  padding-bottom: 30px;
-`;
-const StyledInput = styled.input`
-  border: 1px solid 3e5336;
-  border-radius: 10px;
-  height: 33px;
-  padding-left: 10px;
-  font-size: 20px;
-  text-align: left;
-  margin-top: 5px;
-  margin-bottom: 20px;
-  margin-left: 5px;
-  margin-right: 5px;
-`;
-
-const StyledError = styled.p`
-  color: red;
-  font-size: 18px;
-`;
-
-const StyledButton = styled.button`
-  type: submit;
-  display: inline-block;
-  border: 1px solid 3e5336;
-  border-radius: 10px;
-  font-size: 25px;
-  padding: 0.5rem 0;
-  width: 11rem;
-  height: 50px;
-  background-color: #3e5336;
-  color: white;
-`;
+import {
+  StyledSubmit,
+  StyledCon,
+  StyledForm,
+  StyledInput,
+  StyledError,
+  StyledHead,
+  StyledText,
+} from "./StyledComponents.jsx";
 
 function SignIn() {
   const navigate = useNavigate();
@@ -103,8 +56,8 @@ function SignIn() {
   */
 
   return (
-    <StyledBlock maxWidth="md">
-      <form
+    <StyledCon maxWidth="md">
+      <StyledForm
         onSubmit={(e) => {
           e.preventDefault();
           handleSubmit();
@@ -129,11 +82,11 @@ function SignIn() {
         />
         <br></br>
         <StyledError>{validError}</StyledError>
-        <StyledButton>Login</StyledButton>
+        <StyledSubmit type="submit" value="Login" />
         <br></br>
         <a href="./signup">Don't have an account? Sign up here</a>
-      </form>
-    </StyledBlock>
+      </StyledForm>
+    </StyledCon>
   );
 }
