+++ frontend/src/components/profile/ProfileEdit.jsx
@@ -1,6 +1,15 @@
 import { Button, Paper, TextField } from "@mui/material";
 import React, { useRef, useState } from "react";
 import { useNavigate } from "react-router-dom";
+import {
+  StyledSubmit,
+  StyledCon,
+  StyledForm,
+  StyledInput,
+  StyledError,
+  StyledHead,
+  StyledText,
+} from "../StyledComponents.jsx";
 
 function ProfileEdit() {
   const username = useRef();
@@ -33,14 +42,14 @@ function ProfileEdit() {
     <Paper className="profile" elevation={2}>
       <h1 className="profile-header">&#127820;Edit Profile&#127820;</h1>
       <hr />
-      <form className="edit-form" onSubmit={handleSubmit}>
+      <StyledForm className="edit-form" onSubmit={handleSubmit}>
         <TextField label="Username" inputRef={username} error={error} />
         <TextField label="First Name" inputRef={fname} error={error} />
         <TextField label="Last Name" inputRef={lname} error={error} />
         <Button variant="contained" type="submit">
           Submit
         </Button>
-      </form>
+      </StyledForm>
     </Paper>
   ) : null;
 }
